ref(util): move common functions to util package
* prevents import cycle conflicts in tests
