+++ action/lint.go
@@ -11,6 +11,7 @@ import (
 	"github.com/helm/helm/chart"
 	"github.com/helm/helm/log"
 	"github.com/helm/helm/manifest"
+	helm "github.com/helm/helm/util"
 	"gopkg.in/yaml.v2"
 )
 
@@ -32,7 +33,7 @@ type GHRepoService interface {
 //
 // - homedir is the home directory for the user
 func LintAll(homedir string) {
-	md := filepath.Join(homedir, WorkspaceChartPath, "*")
+	md := filepath.Join(homedir, helm.WorkspaceChartPath, "*")
 	charts, err := filepath.Glob(md)
 	if err != nil {
 		log.Warn("Could not find any charts in %q: %s", md, err)
@@ -55,7 +56,7 @@ func Lint(chartName, homedir string) {
 	var errors = make([]string, 0)
 
 	//assumes chart is in your workspace directory
-	chartPath := filepath.Join(homedir, WorkspaceChartPath, chartName)
+	chartPath := filepath.Join(homedir, helm.WorkspaceChartPath, chartName)
 	//makes sure all files are in place
 	structure, fatalDs := directoryStructure(chartPath)
