+++ action/lint_test.go
@@ -8,39 +8,40 @@ import (
 	"strings"
 	"testing"
 
+	"github.com/helm/helm/test"
+	"github.com/helm/helm/util"
+
 	"gopkg.in/yaml.v2"
 )
 
 func TestLintSuccess(t *testing.T) {
-	tmpHome := createTmpHome()
-	fakeUpdate(tmpHome)
+	tmpHome := test.CreateTmpHome()
+	test.FakeUpdate(tmpHome)
 
 	chartName := "goodChart"
 
 	Create(chartName, tmpHome)
 
-	output := capture(func() {
+	output := test.CaptureOutput(func() {
 		Lint(chartName, tmpHome)
 	})
 
 	expected := "Chart [goodChart] has passed all necessary checks"
 
-	if !strings.Contains(output, expected) {
-		t.Fatalf("Expected: '%s' in %s.", expected, output)
-	}
+	test.ExpectContains(t, output, expected)
 }
 
 func TestLintMissingReadme(t *testing.T) {
-	tmpHome := createTmpHome()
-	fakeUpdate(tmpHome)
+	tmpHome := test.CreateTmpHome()
+	test.FakeUpdate(tmpHome)
 
 	chartName := "badChart"
 
 	Create(chartName, tmpHome)
 
-	os.Remove(filepath.Join(tmpHome, WorkspaceChartPath, chartName, "README.md"))
+	os.Remove(filepath.Join(tmpHome, util.WorkspaceChartPath, chartName, "README.md"))
 
-	output := capture(func() {
+	output := test.CaptureOutput(func() {
 		Lint(chartName, tmpHome)
 	})
 
@@ -52,51 +53,26 @@ func TestLintMissingReadme(t *testing.T) {
 }
 
 func TestLintMissingChartYaml(t *testing.T) {
-	tmpHome := createTmpHome()
-	fakeUpdate(tmpHome)
+	tmpHome := test.CreateTmpHome()
+	test.FakeUpdate(tmpHome)
 
 	chartName := "badChart"
 
 	Create(chartName, tmpHome)
 
-	os.Remove(filepath.Join(tmpHome, WorkspaceChartPath, chartName, "Chart.yaml"))
+	os.Remove(filepath.Join(tmpHome, util.WorkspaceChartPath, chartName, "Chart.yaml"))
 
-	output := capture(func() {
+	output := test.CaptureOutput(func() {
 		Lint(chartName, tmpHome)
 	})
 
-	expectContains(t, output, "A Chart.yaml file was not found")
-	expectContains(t, output, "Chart [badChart] failed some checks")
-}
-
-func TestLintAllNone(t *testing.T) {
-	tmpHome := createTmpHome()
-	fakeUpdate(tmpHome)
-
-	output := capture(func() {
-		Cli().Run([]string{"helm", "--home", tmpHome, "lint", "--all"})
-	})
-
-	expectContains(t, output, fmt.Sprintf("Could not find any charts in \"%s", tmpHome))
-}
-
-func TestLintSingleCli(t *testing.T) {
-	tmpHome := createTmpHome()
-	fakeUpdate(tmpHome)
-
-	chartName := "goodChart"
-	Create(chartName, tmpHome)
-
-	output := capture(func() {
-		Cli().Run([]string{"helm", "--home", tmpHome, "lint", chartName})
-	})
-
-	expectContains(t, output, fmt.Sprintf("Chart [%s] has passed all necessary checks", chartName))
+	test.ExpectContains(t, output, "A Chart.yaml file was not found")
+	test.ExpectContains(t, output, "Chart [badChart] failed some checks")
 }
 
 func TestLintEmptyChartYaml(t *testing.T) {
-	tmpHome := createTmpHome()
-	fakeUpdate(tmpHome)
+	tmpHome := test.CreateTmpHome()
+	test.FakeUpdate(tmpHome)
 
 	chartName := "badChart"
 
@@ -104,38 +80,18 @@ func TestLintEmptyChartYaml(t *testing.T) {
 
 	badChartYaml, _ := yaml.Marshal(make(map[string]string))
 
-	chartYaml := filepath.Join(tmpHome, WorkspaceChartPath, chartName, "Chart.yaml")
+	chartYaml := filepath.Join(tmpHome, util.WorkspaceChartPath, chartName, "Chart.yaml")
 
 	os.Remove(chartYaml)
 	ioutil.WriteFile(chartYaml, badChartYaml, 0644)
 
-	output := capture(func() {
+	output := test.CaptureOutput(func() {
 		Lint(chartName, tmpHome)
 	})
 
-	expectContains(t, output, "Missing Name specification in Chart.yaml file")
-	expectContains(t, output, "Missing Version specification in Chart.yaml file")
-	expectContains(t, output, "Missing description in Chart.yaml file")
-	expectContains(t, output, "Missing maintainers information in Chart.yaml file")
-	expectContains(t, output, fmt.Sprintf("Chart [%s] failed some checks", chartName))
-}
-
-func TestLintAll(t *testing.T) {
-	tmpHome := createTmpHome()
-	fakeUpdate(tmpHome)
-
-	missingReadmeChart := "missingReadme"
-
-	Create(missingReadmeChart, tmpHome)
-	os.Remove(filepath.Join(tmpHome, WorkspaceChartPath, missingReadmeChart, "README.md"))
-
-	Create("goodChart", tmpHome)
-
-	output := capture(func() {
-		Cli().Run([]string{"helm", "--home", tmpHome, "lint", "--all"})
-	})
-
-	expectMatches(t, output, "A README file was not found.*"+missingReadmeChart)
-	expectContains(t, output, "Chart [goodChart] has passed all necessary checks")
-	expectContains(t, output, "Chart [missingReadme] failed some checks")
+	test.ExpectContains(t, output, "Missing Name specification in Chart.yaml file")
+	test.ExpectContains(t, output, "Missing Version specification in Chart.yaml file")
+	test.ExpectContains(t, output, "Missing description in Chart.yaml file")
+	test.ExpectContains(t, output, "Missing maintainers information in Chart.yaml file")
+	test.ExpectContains(t, output, fmt.Sprintf("Chart [%s] failed some checks", chartName))
 }
