+++ action/list.go
@@ -5,11 +5,12 @@ import (
 
 	"github.com/helm/helm/chart"
 	"github.com/helm/helm/log"
+	helm "github.com/helm/helm/util"
 )
 
 // List lists all of the local charts.
 func List(homedir string) {
-	md := filepath.Join(homedir, WorkspaceChartPath, "*")
+	md := filepath.Join(homedir, helm.WorkspaceChartPath, "*")
 	charts, err := filepath.Glob(md)
 	if err != nil {
 		log.Warn("Could not find any charts in %q: %s", md, err)
