+++ helm.go
@@ -3,5 +3,5 @@ package main
 import "github.com/helm/helm/cli"
 
 func main() {
-	cli.Execute()
+	cli.Cli().RunAndExitOnError()
 }
